refactor(tmuxrun): tmuxctl の display-message 系 2 関数を統合
internal/tmuxctl は DisplayMessage / DisplayMessageToClient のみの
極小パッケージで、exec("tmux") ラップが tmuxrun と重複していた。
本体とテストを internal/tmuxrun/display.go へ移送し、呼び出し元
3 箇所 (notify, dashboard, focusconsole) を差し替えて internal/tmuxctl
を削除する。同一パッケージ化に伴い、target 修飾は既存の
exactSessionTarget を再利用し、2 関数の共通部を displayMessage に、
'#' エスケープを tmuxLiteral に一本化した。エラー文言・tmux 引数は不変。
