+++ cmd/fanout/focusconsole.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/butaosuinu/fanout/internal/exitcode"
 	"github.com/butaosuinu/fanout/internal/log"
-	"github.com/butaosuinu/fanout/internal/tmuxctl"
 	"github.com/butaosuinu/fanout/internal/tmuxrun"
 )
 
@@ -73,7 +72,7 @@ func cmdFocusConsole(args []string, lg *log.Logger) exitcode.Code {
 	}
 	console, ok := pickConsolePane(from, panes)
 	if !ok {
-		if err := tmuxctl.DisplayMessageToClient(flags.client, "fanout: no live console; run 'fanout' to start one"); err != nil {
+		if err := tmuxrun.DisplayMessageToClient(flags.client, "fanout: no live console; run 'fanout' to start one"); err != nil {
 			lg.Debug("focus-console: %v", err)
 		}
 		return exitcode.OK
